Revert "seventh  version"
Новая попытка исправления замечаний после ревью кода
