+++ src/dvorenenko/ru/exception/ErrorExtensionException.java
@@ -1,4 +1,4 @@
-package MyException;
+package dvorenenko.ru.exception;
 
 public class ErrorExtensionException extends Exception{
     public ErrorExtensionException(String message){super(message);}
